Move old beatmap download button to more general namespace
