+++ osu.Game/Beatmaps/Drawables/DownloadProgressBar.cs
@@ -5,13 +5,12 @@ using osu.Framework.Allocation;
 using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
-using osu.Game.Beatmaps;
 using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Online;
 using osuTK.Graphics;
 
-namespace osu.Game.Overlays.BeatmapListing.Panels
+namespace osu.Game.Beatmaps.Drawables
 {
     public class DownloadProgressBar : CompositeDrawable
     {
